+++ docs_src/tutorial/plugin.py
@@ -2,7 +2,7 @@ import sys
 
 from pydantic import BaseModel
 from pytauri import AppHandle, Commands
-from pytauri_plugin_notification import NotificationBuilderArgs, NotificationExt
+from pytauri_plugin_notification import NotificationExt
 
 commands: Commands = Commands()
 
@@ -18,9 +18,7 @@ class Greeting(BaseModel):
 @commands.command()
 async def greet(body: Person, app_handle: AppHandle) -> Greeting:
     notification_builder = NotificationExt.builder(app_handle)
-    notification_builder.show(
-        NotificationBuilderArgs(title="Greeting", body=f"Hello, {body.name}!")
-    )
+    notification_builder.show(title="Greeting", body=f"Hello, {body.name}!")
 
     return Greeting(
         message=f"Hello, {body.name}! You've been greeted from Python {sys.version}!"