+++ python/pytauri-plugin-notification/src/pytauri_plugin_notification/__init__.py
@@ -3,13 +3,11 @@
 from pytauri_plugin_notification.ffi import (
     ImplNotificationExt,
     NotificationBuilder,
-    NotificationBuilderArgs,
     NotificationExt,
 )
 
 __all__ = [
     "ImplNotificationExt",
     "NotificationBuilder",
-    "NotificationBuilderArgs",
     "NotificationExt",
 ]