+++ crates/pytauri-plugin-notification/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## [Unreleased]
 
+### BREAKING
+
+- [#70](https://github.com/WSH032/pytauri/pull/70) - Removed `NotificationBuilderArgs`
+
+### Added
+
+- [#70](https://github.com/WSH032/pytauri/pull/70) - feat: add more `NotificationBuilder` parameters
+    - `id`
+    - `channel_id`
+    - `large_body`
+    - `summary`
+    - `action_type_id`
+    - `group`
+    - `group_summary`
+    - `sound`
+    - `inbox_line`
+    - `icon`
+    - `large_icon`
+    - `icon_color`
+    - `ongoing`
+    - `auto_cancel`
+    - `silent`
+
 ### Changed
 
 - [#47](https://github.com/WSH032/pytauri/pull/47) - refactor: use `pytauri::ImplManager` as `self::ImplNotificationExt`